only administrators can change roles

+++ server/repository/public/user/update.ts
@@ -5,12 +5,13 @@ import { BaseProtocol, DB } from '@server/db'
 import { getOne } from './getOne'
 
 export const update = async (
-  props: { user: User; profilePicture?: Express.Multer.File | null },
+  props: { user: User; profilePicture?: Express.Multer.File | null; isAdministrator: boolean },
   client: BaseProtocol = DB
 ): Promise<User> => {
   const {
     profilePicture,
     user: { institution, lang, name, status, position, email, id, roles },
+    isAdministrator,
   } = props
 
   await client.one<User>(
@@ -28,18 +29,20 @@ export const update = async (
     [institution, lang, name, status, position, email, id]
   )
 
-  await client.query(`delete from users_role WHERE user_id = $1`, [id])
+  if (isAdministrator) {
+    await client.query(`delete from users_role WHERE user_id = $1`, [id])
 
-  const userRolePromises = roles
-    .filter((userRole) => !!userRole)
-    .map((userRole: UserRole<RoleName>) =>
-      client.query(
-        `insert into users_role (user_id, assessment_id, cycle_uuid, country_iso, role, accepted_at) values ($1, $2, $3, $4, $5, now())`,
-        [id, userRole.assessmentId, userRole.cycleUuid, userRole.countryIso, userRole.role]
+    const userRolePromises = roles
+      .filter((userRole) => !!userRole)
+      .map((userRole: UserRole<RoleName>) =>
+        client.query(
+          `insert into users_role (user_id, assessment_id, cycle_uuid, country_iso, role, accepted_at) values ($1, $2, $3, $4, $5, now())`,
+          [id, userRole.assessmentId, userRole.cycleUuid, userRole.countryIso, userRole.role]
+        )
       )
-    )
 
-  await Promise.all(userRolePromises)
+    await Promise.all(userRolePromises)
+  }
 
   if (profilePicture) {
     const {
